pyproject.toml
[tool.black]
line-length = 99
target-version = ["py37", "py38", "py39", "py310"]
exclude = '''
/(
    \.git|.venv
)/
'''

[tool.isort]
profile = "black"
float_to_top = true
skip_glob = ".venv"

[tool.poetry]
name = "python-incidentio-client"
version = "0.1.0"
description = "Python client for Incident.io"
repository = "https://github.com/expobrain/python-incidentio-client"
authors = ["Daniele Esposti <daniele.esposti@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "incident_io_client" }]

[tool.poetry.dependencies]
python = "^3.7"
httpx = ">=0.15.4,<0.23.0"
attrs = ">=21.3.0"
python-dateutil = "^2.8.0"

[tool.poetry.dev-dependencies]
autoflake = "^1.4"
black = "^22.1"
isort = "^5.10.1"
mypy = "^0.942"
flake8 = "^4.0.1"
types-python-dateutil = "^2.8.10"
pyupgrade = "^2.31.1"
openapi-python-client = "^0.11.1"
types-certifi = "^2021.10.8"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
exclude = ".venv/"
check_untyped_defs = true
disallow_incomplete_defs = true
warn_unused_configs = true
warn_unused_ignores = true
disallow_subclassing_any = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_return_any = true
